# consent-bb/docs/openapi/v2023.8.2/bundled.yaml
components:
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: Access token passed with Bearer prefix in Authorization header
      scheme: bearer
      type: http
info:
  contact:
    email: support@igrant.io
    name: iGrant.io
    url: 'https://igrant.io'
  description: 'The Consent Building Block enables services for individuals to approve the use  of their personal data by defining the principles, functions, and architecture  of an information system. For organisations that process personal data, it  provides the ability to know the individual''s will and legitimately process such  personal data. The Consent Building Block is a process-oriented GovStack Building Block  facilitating auditable bilateral data-agreements within a multi-agent environment that  integrates with most other Building Blocks.'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
  title: Consent Building Block APIs
  version: 2023.8.2
openapi: 3.1.0
paths:
...
  /onboard/token/refresh:
    post:
      deprecated: false
      description: Refresh access token
      operationId: refreshToken
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                clientId:
                  type: string
                refreshToken:
                  type: string
              required:
                - refreshToken
                - clientId
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/paths/~1onboard~1admin~1login/post/responses/200/content/application~1json/schema'
          description: OK
      security: []
      tags:
        - onboard
...
servers:
  - description: Development server
    url: 'https://api.bb-consent.dev/v2'
  - description: Staging server
    url: 'https://staging-consent-bb-api.igrant.io/v2'
  - description: Demo server
    url: 'https://demo-consent-bb-api.igrant.io/v2'
...